+++ backend/package.json
@@ -19,7 +19,7 @@
     "prisma": "^5.14.0",
     "supertest": "^7.0.0",
     "ts-jest": "^29.1.4",
-    "ts-node": "^10.9.2",
+    "ts-node": "^10.9.1",
     "typescript": "^4.5.2"
   },
   "dependencies": {
