+++ backend/src/libs/nodemailer.ts
@@ -5,7 +5,7 @@ export const transporter = nodemailer.createTransport({
   port: 587,
   secure: false, // Use `true` for port 465, `false` for all other ports
   auth: {
-    user: "bagushendrawan892@gmail.com",
-    pass: "wpiz dsgd otlq hgeq", // application password / app password
+    user: process.env.EMAIL_USER,
+    pass: process.env.EMAIL_PASSWORD, // application password / app password
   },
 });
