+++ frontend/src/features/profilePage/entities/edit-profile.ts
@@ -1,0 +1,8 @@
+type editProfileForm = {
+    photoProfile?: string,
+    fullName : string,
+    username : string,
+    bio : string,
+    followersCount: number,
+    followingsCount: number
+}
